requirements.txt
torch==2.0.1
torchvision>=0.15.2
lightning==2.0.0
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

omegaconf
rootutils
wandb
tensorboard

diffusers==0.20.0
transformers==4.28.1
xformers==0.0.22
accelerate==0.23.0

einops==0.6.1
opencv-python==4.7.0.72
protobuf==3.20.3
wrapt==1.15.0
numpy==1.23.5
google==3.0.0
googleapis-common-protos==1.59.0
timm==0.9.7
kornia==0.7.0
matplotlib
scikit-learn
imageio

# for eval mesh
open3d
mesh2sdf
transforms3d
git+https://github.com/NVlabs/nvdiffrast.git
